+++ src/onegov.org/onegov/org/upgrade.py
@@ -49,3 +49,24 @@ def add_image_dimensions_to_html(context):
 
     for form in context.session.query(FormDefinition):
         form.text = annotate_html(form.text, context.request)
+
+
+@upgrade_task('Remove official notices table')
+def remove_official_notices_table(context):
+    # an incompatible release was accidentally left in, so we remove the
+    # table in such instances, triggering a recreation on the next start
+
+    if not context.has_table("official_notices"):
+        return False
+
+    session = context.app.session_manager.session()
+
+    organisations_count = session.execute("select count(*) from organisations")
+    if organisations_count.scalar() != 1:
+        return False
+
+    notices_count = session.execute("select count(*) from official_notices")
+    if notices_count.scalar() != 0:
+        return False
+
+    context.operations.drop_table("official_notices")